Share FailingRulesConverter
Move FailingRulesConverter to shared utils
Extend FailingRulesConverter with kwargs support
Specify list of params for failing rule
